fixed issues and styled
