+++ front-end/src/app/category/category.component.html
@@ -1,32 +1,45 @@
-<div><h1>Category</h1></div>
-
-  <div>
-    <button (click)="goToCategory(1)">Comedy</button>
-
+<div class="col-md-4 text-center"><h1>Category</h1></div>
+<br/>
+<br />
+<div class="row form-group">
+  <div class=" col-md-4 text-center">
+    <button class="btn btn-primary btn-lg btn-block" (click)="goToCategory(1)">Comedy</button>
   </div>
+</div>
+<br />
 
-  <div>
-    <button (click)="goToCategory(2)">Action</button>
-
+<div class="row form-group">
+  <div class=" col-md-4 text-center">
+    <button class="btn btn-primary btn-lg btn-block" (click)="goToCategory(2)">Action</button>
   </div>
+</div>
+<br />
 
-  <div>
-    <button (click)="goToCategory(3)">Horror</button>
-
+<div class="row ">
+  <div class=" col-md-4 text-center form-group">
+    <button class="btn btn-primary btn-lg btn-block" (click)="goToCategory(3)">Horror</button>
   </div>
+</div>
+<br />
 
-  <div>
-    <button (click)="goToCategory(4)">Romance</button>
-
+<div class="row form-group">
+  <div class=" col-md-4 text-center">
+    <button class="btn btn-primary btn-lg btn-block" (click)="goToCategory(4)">Romance</button>
   </div>
+</div>
+<br />
 
-  <div>
-    <button (click)="goToCategory(5)">Sci-Fi</button>
-
+<div class="row form-group">
+  <div class=" col-md-4 text-center">
+    <button class="btn btn-primary btn-lg btn-block" (click)="goToCategory(5)">Sci-Fi</button>
   </div>
+</div>
+<br />
 
-  <div>
-    <button (click)="goToCategory(6)">Thriller</button>
-
+<div class="row form-group">
+  <div class=" col-md-4 text-center">
+    <button class="btn btn-primary btn-lg btn-block" (click)="goToCategory(6)">Thriller</button>
   </div>
+</div>
+<br />
