+++ front-end/src/app/category/category.component.ts
@@ -14,9 +14,8 @@ export class CategoryComponent implements OnInit {
   }
 
   goToCategory(category: number) {
-    this.router.navigate(['/question', category]);
+    this.router.navigate(['/question', category], { skipLocationChange: false });
     console.log(category);
   }
-  
   
 }
