+++ front-end/src/app/questions/questions.component.css
@@ -1,7 +1,3 @@
-<<<<<<< HEAD
-=======
-
->>>>>>> dev
 .funkyradio div {
   clear: both;
   overflow: hidden;
