--- front-end/src/app/shared/user.model.spec.ts
@@ -1,7 +1,0 @@
-import { User } from './user.model';
-
-describe('User', () => {
-  it('should create an instance', () => {
-    expect(new User()).toBeTruthy();
-  });
-});
