--- front-end/src/app/shared/user.model.ts
@@ -1,5 +1,0 @@
-export class User {
-  fullName: string;
-    email: string;
-    password: string;
-}
