--- front-end/src/app/shared/user.service.ts
@@ -1,65 +1,0 @@
-import { Injectable } from '@angular/core';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
-
-import { environment } from '../../environments/environment';
-
-import { User } from './user.model';
-import { Subject } from 'rxjs';
-@Injectable({
-  providedIn: 'root'
-})
-export class UserService {
-
-
-  selectedUser: User = {
-    fullName: '',
-    email: '',
-    password: ''
-  };
-  noAuthHeader = { headers: new HttpHeaders({ 'NoAuth': 'True' }) };
-
-  constructor(private http: HttpClient) { }
-
-  postUser(user: User){
-    return this.http.post(environment.apiBaseUrl+'/register',user,this.noAuthHeader)
-  }
-
-  
-  /*addPost(post: Post){
-    return this.http.post('http://localhost:3000/api/post/createPost',{
-        title : post.title,
-        description : post.description
-    })
-}*/
-  login(authCredentials) {
-    return this.http.post(environment.apiBaseUrl + '/authenticate', authCredentials,this.noAuthHeader);
-  }
-  getUserProfile() {
-    return this.http.get(environment.apiBaseUrl + '/category');
-  }
-  setToken(token: string) {
-    localStorage.setItem('token', token);
-  }
-  getToken() {
-    return localStorage.getItem('token');
-  }
-  deleteToken() {
-    localStorage.removeItem('token');
-  }
-  getUserPayload() {
-    var token = this.getToken();
-    if (token) {
-      var userPayload = atob(token.split('.')[1]);
-      return JSON.parse(userPayload);
-    }
-    else
-      return null;
-  }
-  isLoggedIn() {
-    var userPayload = this.getUserPayload();
-    if (userPayload)
-      return userPayload.exp > Date.now() / 1000;
-    else
-      return false;
-  }
-}
