+++ front-end/src/app/user/sign-in/sign-in.component.spec.ts
@@ -1,6 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { SignInComponent } from './sign-in.component';
+import { RouterTestingModule } from '@angular/router/testing';
+import { HttpClientModule } from '@angular/common/http';
+import { UserService } from '../../services/user.service';
+import { FormsModule } from '@angular/forms';
 
 describe('SignInComponent', () => {
   let component: SignInComponent;
@@ -8,6 +12,10 @@ describe('SignInComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
+      imports: [RouterTestingModule, HttpClientModule, FormsModule],
+      providers: [
+        UserService
+      ],
       declarations: [ SignInComponent ]
     })
     .compileComponents();
